Moved rule send to config
